Added a bit of documentation Extended entrypoint to validate if LocalSettings.php was created Added some output to determine at which step the entrypoint is right now

+++ bluespice-entrypoint.sh
@@ -16,18 +16,28 @@ if [ ! -f /config/LocalSettings.php ]; then
                                                       ${ADMIN_LOGIN}
 fi
 
-settings_contains_bluespice_require=$( cat /config/LocalSettings.php | grep 'LocalSettings.BlueSpice.php' | wc -l )
+if [ -f /config/LocalSettings.php ]; then
+        settings_contains_bluespice_require=$( cat /config/LocalSettings.php | grep 'LocalSettings.BlueSpice.php' | wc -l )
 
-if [ "${settings_contains_bluespice_require}" -ne "1" ]; then
-        echo 'require_once "$IP/LocalSettings.BlueSpice.php";' >> /config/LocalSettings.php
+        if [ "${settings_contains_bluespice_require}" -ne "1" ]; then
+                echo "BlueSpice include is missing in LocalSettings.php. Adding it for you..."
+                echo 'require_once "$IP/LocalSettings.BlueSpice.php";' >> /config/LocalSettings.php
+        fi
+
+        echo "Copying persistent LocalSettings.php to web directory..."
+        cp -f /config/LocalSettings.php /var/www/html/
+
+        echo "Copying external extensions to web directory..."
+        cp -rf /extensions/* /var/www/html/extensions/
+
+        echo "Run update script to ensure that database is up to date..."
+        php maintenance/update.php --quick
+
+        echo "Correct permissions to ensure that web server is able to access the web directory..."
+        chown -R www-data:www-data /var/www/html
+
+        echo "Starting Apache. Enjoy your BlueSpice wiki!"
+        apache2-foreground
+elif
+        echo "Setup failed exiting now!"
 fi
-
-cp -f /config/LocalSettings.php /var/www/html/
-
-cp -rf /extensions/* /var/www/html/extensions/
-
-php maintenance/update.php --quick
-
-chown -R www-data:www-data /var/www/html
-
-apache2-foreground
